Creating Product List

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Category;
 use Session;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -15,8 +16,12 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $categories = Category::orderBy('name', 'ASC')->paginate(10);
-        return view('categories.index')->with('categories', $categories);
+        if (Auth::check()) {
+            $categories = Category::orderBy('name', 'ASC')->paginate(10);
+            return view('categories.index')->with('categories', $categories);
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -26,7 +31,11 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view('categories.create');
+        if (Auth::check()) {
+            return view('categories.create');
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -37,19 +46,23 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        //validate the data
-        // if fails, defaults to create() passing errors
-        $this->validate($request, ['name' => 'required|max:100|unique:categories,name']);
+        if (Auth::check()) {
+            //validate the data
+            // if fails, defaults to create() passing errors
+            $this->validate($request, ['name' => 'required|max:100|unique:categories,name']);
 
-        //send to DB (use ELOQUENT)
-        $category = new Category;
-        $category->name = $request->name;
-        $category->save(); //saves to DB
+            //send to DB (use ELOQUENT)
+            $category = new Category;
+            $category->name = $request->name;
+            $category->save(); //saves to DB
 
-        Session::flash('success', 'The category has been added');
+            Session::flash('success', 'The category has been added');
 
-        //redirect
-        return redirect()->route('categories.index');
+            //redirect
+            return redirect()->route('categories.index');
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -71,8 +84,12 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
-        $category = Category::find($id);
-        return view('categories.edit')->with('category', $category);
+        if (Auth::check()) {
+            $category = Category::find($id);
+            return view('categories.edit')->with('category', $category);
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -84,20 +101,24 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //validate the data
-        // if fails, defaults to create() passing errors
-        $category = Category::find($id);
-        $this->validate($request, ['name' => "required|max:100|unique:categories,name,$id"]);
+        if (Auth::check()) {
+            //validate the data
+            // if fails, defaults to create() passing errors
+            $category = Category::find($id);
+            $this->validate($request, ['name' => "required|max:100|unique:categories,name,$id"]);
 
-        //send to DB (use ELOQUENT)
-        $category->name = $request->name;
+            //send to DB (use ELOQUENT)
+            $category->name = $request->name;
 
-        $category->save(); //saves to DB
+            $category->save(); //saves to DB
 
-        Session::flash('success', 'The category has been updated');
+            Session::flash('success', 'The category has been updated');
 
-        //redirect
-        return redirect()->route('categories.index');
+            //redirect
+            return redirect()->route('categories.index');
+        } else {
+            return view('auth.login');
+        }
     }
 
     /**
@@ -108,14 +129,18 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category = Category::find($id);
-        $result = Category::find($id)->items;
-        if ($result->isEmpty()) {
-            $category->delete();
-            Session::flash('success', 'The category has been deleted');
+        if (Auth::check()) {
+            $category = Category::find($id);
+            $result = Category::find($id)->items;
+            if ($result->isEmpty()) {
+                $category->delete();
+                Session::flash('success', 'The category has been deleted');
+            }
+
+
+            return redirect()->route('categories.index');
+        } else {
+            return view('auth.login');
         }
-
-
-        return redirect()->route('categories.index');
     }
 }
